Cleaned Code
Made Bar at home page own File
